dvc: commit experiment 6fd9ae4da944d59ddfb9a431c693d1d5e01645f920ec19e5bc32b8ba50cdcadf

+++ train.py
@@ -6,6 +6,7 @@ import torch
 import torch.nn.functional as F
 import torchvision
 import dvclive
+import numpy as np
 
 
 EPOCHS = 10
@@ -100,7 +101,7 @@ def main():
             # Evaluate and checkpoint.
             metrics = evaluate(model, x_test, y_test)
             for metric, value in metrics.items():
-                dvclive.log(metric, value)
+                dvclive.log(metric, np.float(value))
             dvclive.next_step()
     except KeyboardInterrupt:
         pass
